requirements.txt
torch
torchvision
numpy
einops
thop
timm
opencv-python
scikit-image
pillow
hdf5storage
ninja
lmdb
requests
tqdm
matplotlib
scipy
tensorboard
pyyaml
